// src/utils/enums/TiposDePokemonEnum.ts
export enum TiposDePokemonEnum {
    BUG = "bug",
    DARK = "dark",
    DRAGON = "dragon",
    ELECTRIC = "electric",
    FAIRY = "fairy",
    FIGHTING = "fighting",
    FIRE = "fire",
    FLYING = "flying",
    GHOST = "ghost",
    GRASS = "grass",
    GROUND = "ground",
    ICE = "ice",
    NORMAL = "normal",
    POISON = "poison",
    PSYCHIC = "psychic",
    ROCK = "rock",
    STEEL = "steel",
    WATER = "water"
}